ticket(0458): note potential sites instead of aliasing (no specific correspondence)
Per author: don't put the names in the Project alias column — that falsely
asserts a 1:1 match to a specific NĐ Miền Bắc slot. The three E542 PL9.2
candidate locations have no specific correspondence to any one cancelled slot.

- Replace add_aliases_0458.py with note_potential_sites_0458.py: appends a
  free-text Note on NĐ Miền Bắc 1/2/3 stating the non-correspondence and citing
  E542 PL9.2, preserving NĐ Miền Bắc 1's existing note. Project alias column
  left empty.
- Regenerate pipeline+0458-2026-06-09.ods (Note column is non-extracted →
  reference still byte-identical at 177; convergence gate green).
- config / PROVENANCE / MASTER_REVERSE_SYNC_0458.md / ticket close text updated
  from "aliases" to "notes (no specific correspondence)".
